+++ Data-Visualisation/test.txt
@@ -1,0 +1,5 @@
+1 2 4 4 5 6 7 1 9 10
+2 6 3 4 5 6 7 9 9 8
+1 5 5 4 5 6 7 7 9 5
+5 8 10 4 5 6 7 10 9 7
+4 11 12 4 5 6 7 10 9 4
