+++ Data-Visualisation/write_simple_txt/data_struct.cpp
@@ -1,0 +1,13 @@
+/*#include "data_struct.h"
+
+int data_struct::reset( void )
+ {
+    this->temperature=0;
+    this->height=0;
+    for(int i=0;i<3;i++){
+        this->acceleration[i]=0;
+        this->orientation[i]=0;
+    }
+    return 0;   
+}
+//*/
