+++ Designs/engine/info.txt
@@ -1,0 +1,19 @@
+Επιθυμούμε την κατασκευή ακροφυσίου (nozzle) το οποίο θα ενωθεί με τον θάλαμο
+καύσης (chamber). Το υλικό του ακροφυσίου επιθυμούμε να είναι αλουμίνιο 
+οποιασδήποτε κραματικής σειράς. Ο σωλήνας (θάλαμος καύσης) βρίσκεται στην
+κατοχή μας και μπορεί να προσκομιστεί στο μηχανουργείο.
+
+Η πρώτη σκέψη ήταν να ενωθεί με σπείρες (θηλυκό σπείρωμα στο
+chamber και αρσενικό στο nozzle). Ωστόσο, αυτό δεν γνωρίζω κατά πόσο είναι 
+εφικτό διότι και τα δύο κομμάτια έχουν μικρό πάχος και μπορεί να παρουσιαστούν
+δομικά προβλήματα. Επίσης μπορεί να μην είναι διαθέσιμα τα απαιτούμενα εργαλεία
+στο μηχανουργείο (κοπτικά, κολαούζα, φιλιέρες). Αν αυτό όμως είναι δυνατόν να 
+κατασκευαστεί, δεν υπάρχει κάποια ιδιάιτερη προτίμηση για τον τύπο και τις 
+διαστάσεις των σπειρών.
+
+Η δεύτερη σκέψη ήταν να ενωθεί με δύο ακέφαλες βίδες (σκουλήκια) ώστε να 
+συγκρατούνται τα δύο κομμάτια. Ο λόγος που δεν θέλουμε να συγκολληθούν τα δύο
+κομμάτια είναι διότι χρειάζεται να τοποθετείται πριν κάθε χρήση ο κινητήρας 
+καυσίμου στο chamber.
+
+Τα σχέδια για κάθε περίπτωση βρίσκονται στους δύο υποφακέλους.
