+++ Designs/fin-assembly/STL/fin.stl
@@ -1,0 +1,198 @@
+solid CATIA STL
+  facet normal  1.000000e+000  -0.000000e+000  0.000000e+000
+    outer loop
+      vertex  3.300000e+000  8.600000e+001  8.600000e+001
+      vertex  3.300000e+000  4.300000e+001  8.600000e+001
+      vertex  3.300000e+000  0.000000e+000  0.000000e+000
+    endloop
+  endfacet
+  facet normal  1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  3.300000e+000  8.600000e+001  8.600000e+001
+      vertex  3.300000e+000  0.000000e+000  0.000000e+000
+      vertex  3.300000e+000  2.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  3.300000e+000  8.600000e+001  8.600000e+001
+      vertex  3.300000e+000  2.800000e+001  0.000000e+000
+      vertex  3.300000e+000  5.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  3.300000e+000  8.600000e+001  8.600000e+001
+      vertex  3.300000e+000  5.800000e+001  0.000000e+000
+      vertex  3.300000e+000  8.600000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  3.300000e+000  2.800000e+001 -2.100000e+000
+      vertex  3.300000e+000  5.800000e+001 -2.100000e+000
+      vertex  3.300000e+000  2.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  1.000000e+000  -0.000000e+000  0.000000e+000
+    outer loop
+      vertex  3.300000e+000  2.800000e+001  0.000000e+000
+      vertex  3.300000e+000  5.800000e+001 -2.100000e+000
+      vertex  3.300000e+000  5.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal -1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  8.600000e+001  8.600000e+001
+      vertex  0.000000e+000  8.600000e+001  0.000000e+000
+      vertex  0.000000e+000  5.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal -1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  8.600000e+001  8.600000e+001
+      vertex  0.000000e+000  5.800000e+001  0.000000e+000
+      vertex  0.000000e+000  2.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal -1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  8.600000e+001  8.600000e+001
+      vertex  0.000000e+000  2.800000e+001  0.000000e+000
+      vertex  0.000000e+000  0.000000e+000  0.000000e+000
+    endloop
+  endfacet
+  facet normal -1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  8.600000e+001  8.600000e+001
+      vertex  0.000000e+000  0.000000e+000  0.000000e+000
+      vertex  0.000000e+000  4.300000e+001  8.600000e+001
+    endloop
+  endfacet
+  facet normal -1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  2.800000e+001 -2.100000e+000
+      vertex  0.000000e+000  2.800000e+001  0.000000e+000
+      vertex  0.000000e+000  5.800000e+001 -2.100000e+000
+    endloop
+  endfacet
+  facet normal -1.000000e+000  0.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  5.800000e+001 -2.100000e+000
+      vertex  0.000000e+000  2.800000e+001  0.000000e+000
+      vertex  0.000000e+000  5.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000  0.000000e+000 -1.000000e+000
+    outer loop
+      vertex  3.300000e+000  8.600000e+001  0.000000e+000
+      vertex  3.300000e+000  5.800000e+001  0.000000e+000
+      vertex  0.000000e+000  8.600000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000  0.000000e+000 -1.000000e+000
+    outer loop
+      vertex  0.000000e+000  8.600000e+001  0.000000e+000
+      vertex  3.300000e+000  5.800000e+001  0.000000e+000
+      vertex  0.000000e+000  5.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000  1.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  5.800000e+001  0.000000e+000
+      vertex  3.300000e+000  5.800000e+001  0.000000e+000
+      vertex  0.000000e+000  5.800000e+001 -2.100000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000  1.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  5.800000e+001 -2.100000e+000
+      vertex  3.300000e+000  5.800000e+001  0.000000e+000
+      vertex  3.300000e+000  5.800000e+001 -2.100000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000  0.000000e+000 -1.000000e+000
+    outer loop
+      vertex  3.300000e+000  5.800000e+001 -2.100000e+000
+      vertex  3.300000e+000  2.800000e+001 -2.100000e+000
+      vertex  0.000000e+000  5.800000e+001 -2.100000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000  0.000000e+000 -1.000000e+000
+    outer loop
+      vertex  0.000000e+000  5.800000e+001 -2.100000e+000
+      vertex  3.300000e+000  2.800000e+001 -2.100000e+000
+      vertex  0.000000e+000  2.800000e+001 -2.100000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000 -1.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  2.800000e+001 -2.100000e+000
+      vertex  3.300000e+000  2.800000e+001 -2.100000e+000
+      vertex  0.000000e+000  2.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000 -1.000000e+000  0.000000e+000
+    outer loop
+      vertex  0.000000e+000  2.800000e+001  0.000000e+000
+      vertex  3.300000e+000  2.800000e+001 -2.100000e+000
+      vertex  3.300000e+000  2.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000  0.000000e+000 -1.000000e+000
+    outer loop
+      vertex  3.300000e+000  2.800000e+001  0.000000e+000
+      vertex  3.300000e+000  0.000000e+000  0.000000e+000
+      vertex  0.000000e+000  2.800000e+001  0.000000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000  0.000000e+000 -1.000000e+000
+    outer loop
+      vertex  0.000000e+000  2.800000e+001  0.000000e+000
+      vertex  3.300000e+000  0.000000e+000  0.000000e+000
+      vertex  0.000000e+000  0.000000e+000  0.000000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000 -8.944272e-001  4.472136e-001
+    outer loop
+      vertex  3.300000e+000  4.300000e+001  8.600000e+001
+      vertex  0.000000e+000  4.300000e+001  8.600000e+001
+      vertex  3.300000e+000  0.000000e+000  0.000000e+000
+    endloop
+  endfacet
+  facet normal  0.000000e+000 -8.944272e-001  4.472136e-001
+    outer loop
+      vertex  3.300000e+000  0.000000e+000  0.000000e+000
+      vertex  0.000000e+000  4.300000e+001  8.600000e+001
+      vertex  0.000000e+000  0.000000e+000  0.000000e+000
+    endloop
+  endfacet
+  facet normal  -0.000000e+000  0.000000e+000  1.000000e+000
+    outer loop
+      vertex  3.300000e+000  8.600000e+001  8.600000e+001
+      vertex  0.000000e+000  8.600000e+001  8.600000e+001
+      vertex  3.300000e+000  4.300000e+001  8.600000e+001
+    endloop
+  endfacet
+  facet normal  0.000000e+000  0.000000e+000  1.000000e+000
+    outer loop
+      vertex  3.300000e+000  4.300000e+001  8.600000e+001
+      vertex  0.000000e+000  8.600000e+001  8.600000e+001
+      vertex  0.000000e+000  4.300000e+001  8.600000e+001
+    endloop
+  endfacet
+  facet normal  0.000000e+000  1.000000e+000  0.000000e+000
+    outer loop
+      vertex  3.300000e+000  8.600000e+001  0.000000e+000
+      vertex  0.000000e+000  8.600000e+001  0.000000e+000
+      vertex  3.300000e+000  8.600000e+001  8.600000e+001
+    endloop
+  endfacet
+  facet normal  0.000000e+000  1.000000e+000  0.000000e+000
+    outer loop
+      vertex  3.300000e+000  8.600000e+001  8.600000e+001
+      vertex  0.000000e+000  8.600000e+001  0.000000e+000
+      vertex  0.000000e+000  8.600000e+001  8.600000e+001
+    endloop
+  endfacet
+endsolid CATIA STL
